requirements.txt
flask==2.3.3
openai-whisper==20230918
numpy==1.24.3
torch==2.0.1
